requirements/test.txt
-r base.txt

pytest~=7.1.0
pytest-xdist~=2.5.0
tox~=3.25.1
black~=22.3.0
coverage~=6.3.2
pytest-cov~=3.0.0
pylint~=2.12.2
pylint-pydantic~=0.1.4
pylint-pytest~=1.1.2
mypy~=0.961
